backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keepiq-backend"
version = "0.1.0"
description = "KeepiQ Telegram bot, API and worker"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.15.0",
  "alembic>=1.14.0",
  "apscheduler>=3.10.4",
  "asyncpg>=0.29.0",
  "dateparser>=1.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "openai>=1.58.0",
  "pydantic-settings>=2.7.0",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.36",
  "structlog>=24.4.0",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.20.0",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.24.0"
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../tests"]
